pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = ["juancamilosuarez3 <juancamilosuarez3@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.6"
django-cors-headers = "3.13.0"
djangorestframework = "3.14.0"
django-rest-framework = "0.1.0"
djoser = "2.1.0"
pillow = "9.4.0"
stripe = "5.1.1"
python-dotenv = "0.21.0"
psycopg2 = "^2.9.5"
drf-yasg = "1.21.5"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"


[tool.poetry.group.dev.dependencies]
black = "23.1.0"
click = "^8.1.3"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
